lambdas/requirements.txt
boto3
boto3-stubs[s3,dynamodb]
python-jose[cryptography]
python-dotenv
requests
